# perf: Batch nested embed parquet decoding