--- a/crates/polars-parquet/src/arrow/read/deserialize/struct_.rs
+++ b/crates/polars-parquet/src/arrow/read/deserialize/struct_.rs
@@ -45,7 +45,7 @@ impl<'a> Iterator for StructIterator<'a> {
             }
         }
         let mut nested = nested.pop().unwrap();
-        let (_, validity) = nested.nested.pop().unwrap().take();
+        let (_, validity) = nested.pop().unwrap();
 
         Some(Ok((
             nested,