This continues on #17542 and now also batches the actual Nesting, not only the top decoder.

I did the same benchmark with `POLARS_MAX_THREADS=1` and we again see a small speedup. But I suspect that the speedup is larger for `FixedSizeList`s and `Struct`s.

```
Benchmark 1: After Optimization
  Time (mean ± σ):     13.306 s ±  0.018 s    [User: 12.661 s, System: 0.614 s]
  Range (min … max):   13.263 s … 13.335 s    10 runs

Benchmark 2: Before Optimization
  Time (mean ± σ):     14.973 s ±  0.065 s    [User: 14.422 s, System: 0.517 s]
  Range (min … max):   14.887 s … 15.129 s    10 runs

Summary
  After Optimization ran
    1.13 ± 0.01 times faster than Before Optimization
```